fix(backend): fix consistency with documentation
