+++ backend/src/main/java/it/greentrails/backend/gestioneattivita/service/ValoriEcosostenibilitaService.java
@@ -1,44 +1,12 @@
 package it.greentrails.backend.gestioneattivita.service;
 
 import it.greentrails.backend.entities.ValoriEcosostenibilita;
-import it.greentrails.backend.gestioneattivita.repository.ValoriEcosostenibilitaRepository;
-import java.util.Optional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
-@Service
-@RequiredArgsConstructor
-public class ValoriEcosostenibilitaService {
+public interface ValoriEcosostenibilitaService {
 
-  private final ValoriEcosostenibilitaRepository repository;
+  ValoriEcosostenibilita saveValori(ValoriEcosostenibilita valori) throws Exception;
 
-  public ValoriEcosostenibilita saveValori(ValoriEcosostenibilita valori) throws Exception {
-    if (valori == null) {
-      throw new Exception("Non è possibile salvare questo valore di ecosostenibilità.");
-    }
-    return repository.save(valori);
-  }
+  boolean deleteValori(ValoriEcosostenibilita valori) throws Exception;
 
-
-  public boolean deleteValori(ValoriEcosostenibilita valori) throws Exception {
-    if (valori == null) {
-      throw new Exception("Non è possibile cancellare questo valore di ecosostenibilità.");
-    }
-    repository.delete(valori);
-    repository.flush();
-    return repository.findById(valori.getId()).isEmpty();
-  }
-
-  public ValoriEcosostenibilita findById(Long id) throws Exception {
-    if (id == null || id < 0) {
-      throw new Exception("L'id non è valido.");
-    }
-    Optional<ValoriEcosostenibilita> valori = repository.findById(id);
-    if (valori.isEmpty()) {
-      throw new Exception("I valori non sono stati trovati.");
-    }
-    return valori.get();
-  }
-
-
+  ValoriEcosostenibilita findById(Long id) throws Exception;
 }
