+++ backend/src/main/java/it/greentrails/backend/gestioneitinerari/adapter/ItinerariAdapter.java
@@ -4,7 +4,7 @@ import it.greentrails.backend.entities.Itinerario;
 import it.greentrails.backend.entities.Preferenze;
 
 // TODO: inserire integrazione modulo AI
-public class GestioneItinerariAdapter {
+public class ItinerariAdapter {
 
   public Itinerario pianificazioneAutomatica(Preferenze preferenze) {
     Itinerario itinerario = new Itinerario();
